// apps/web/src/contracts/generated/sheet/PaperChoice.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Orientation } from "./Orientation";
import type { Paper } from "./Paper";

/**
 * A paper a template suits.
 */
export type PaperChoice = { paper: Paper, orientation: Orientation, };